pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
# See https://setuptools.pypa.io/en/latest/userguide/quickstart.html for more project configuration options.
name = "cosmicfishpie"
dynamic = ["version"]
readme = "README.md"
classifiers = [
    "Intended Audience :: Science/Research",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Cosmology",
]
authors = [
    {name = "Santiago Casas, Matteo Martinelli, Sefa Pamuk, Sabarish V.M.", email = "s.casas@protonmail.com"}
]
requires-python = ">=3.10,<3.13"
license = {file = "LICENSE"}

dependencies = [
    "astropy==6.1.7",
    "camb==1.6.4",
    "Cython==3.1.4",
    "getdist==1.7.3",
    "h5py==3.14.0",
    "ipython==8.37.0",
    "joblib==1.5.2",
    "jupyter==1.1.1",
    "matplotlib==3.10.7",
    "nautilus-sampler==1.0.5",
    "chainconsumer>=1.2.0,<2.0.0",
    "numpy",
    "pandas",
    "PyYAML==6.0.3",
    "scikit-learn==1.7.2",
    "SciPy>=1.8.0",
    "seaborn",
    "colossus",
    "symbolic-pofk@git+https://github.com/DeaglanBartlett/symbolic_pofk.git",
]

[project.urls]
Homepage = "https://github.com/santiagocasas/cosmicfishpie"
Repository = "https://github.com/santiagocasas/cosmicfishpie"
Changelog = "https://github.com/santiagocasas/cosmicfishpie/blob/main/CHANGELOG.md"
# Documentation = "https://cosmicfishpie.readthedocs.io/"

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy>=1.0,<1.20",
    "black>=24.4.2,<25.0",
    "isort>=5.12,<7.1",
    "pytest",
    "pytest-sphinx",
    "pytest-cov",
    "twine>=1.11.0",
    "build",
    "setuptools",
    "wheel",
    "Sphinx>=4.3.0,<8.2.0",
    "furo==2024.1.29",
    "myst-parser>=1.0,<4.1",
    "myst_parser",
    "sphinx-copybutton==0.5.2",
    "sphinx-autobuild==2024.10.3",
    "sphinx-autodoc-typehints==1.23.3",
    "sphinx_rtd_theme>=1.0",
    "packaging",
]

[tool.setuptools.packages.find]
exclude = [
    "*.tests",
    "*.tests.*",
    "tests.*",
    "tests",
    "docs*",
    "scripts*"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
cosmicfishpie = [
    "py.typed",
    "configs/default_boltzmann_yaml_files/*.md",
    "configs/default_boltzmann_yaml_files/*/*.yaml",
    "configs/default_survey_specifications/*.yaml",
    "configs/other_survey_specifications/*.yaml",
    "configs/external_data/*.dat",
]

[tool.setuptools.dynamic]
version = {attr = "cosmicfishpie.version.VERSION"}

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
(
      __pycache__
    | \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.vscode
    | \.venv
    | \bdist\b
    | \bdoc\b
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3

# You can override these pyright settings by adding a personal pyrightconfig.json file.
[tool.pyright]
reportPrivateImportUsage = false

[tool.ruff]
line-length = 115
target-version = "py310"
exclude = ["**/*.ipynb"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
ignore_missing_imports = true
no_site_packages = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
strict_optional = false

[tool.pytest.ini_options]
testpaths = "tests/"
python_classes = [
  "Test*",
  "*Test"
]
log_format = "%(asctime)s - %(levelname)s - %(name)s - %(message)s"
log_level = "DEBUG"
